requirements.txt
torch>=2.1
PyYAML>=6.0
matplotlib>=3.7

